add auth

+++ app/components/Delete.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import axios from 'axios';
+import { useRouter } from 'next/router';
+import ClearIcon from '@mui/icons-material/Clear';
 
 interface DeleteButtonProps {
   tableName: string;
@@ -7,12 +9,14 @@ interface DeleteButtonProps {
   id: number;
   endpointUrl: string;
 }
-
 const DeleteButton: React.FC<DeleteButtonProps> = ({ tableName, primaryColumnName, id, endpointUrl }) => {
+  const router = useRouter()
   const handleDelete = async () => {
     try {
+      const confirmed = window.confirm('Are you sure you want to delete this item?');
+      if (!confirmed){return}
       const response = await axios.delete(`${endpointUrl}/${id}`);
-      console.log(response.data); // 削除成功時のレスポンスデータを表示（任意）
+      router.push(`/users`);
       // ここで必要な追加の処理を実行することもできます
     } catch (error) {
       console.error(error); // エラーハンドリング
@@ -21,9 +25,10 @@ const DeleteButton: React.FC<DeleteButtonProps> = ({ tableName, primaryColumnNam
   };
 
   return (
-    <button onClick={handleDelete}>
-      Delete {tableName} {primaryColumnName}: {id}
-    </button>
+    <div>
+
+<ClearIcon className='icon' onClick={handleDelete}/>
+    </div>
   );
 };
